VideoMediaView. reuse QProcess.waitForFinished()
Sometimes finished_signal is not emitted, looks like the stopping step
is not right. After issuing 'quit', calling waitForFinished always returns
`False` regardless the value of milliseconds passed, eg. 10000 will blocks
for 10 seconds and return `False` :indicating that the operation was
timeout, error occurred, or the process has finished (but the finished
signal was never emitted by the process).