Be aware of empty region.
`IndexError: list index out of range` occurs when there is no media in a
region and the region is due to play its content.
This usually happens when editing a region so that there is no media left in
the region, or when creating new region which by default contains no media.